refactor: move user daemon specific functions to listener-user
Adding the devfsevents daemon code will re-use the infrastructure in
listener.c, so pull out portions that are focused on the per-user
service.
